requirements.txt
Django==5.1.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
Pillow==10.2.0
psycopg2-binary==2.9.9  # Pour PostgreSQL (production)
gunicorn==21.2.0  # Pour le déploiement en production
whitenoise==6.6.0  # Pour servir les fichiers statiques en production
